revised WY inputs for import to WaDE
Except for the 'aggregatedamounts.csv', the other tables successfully import to WaDE QA DB.  Here is the summary of the modifications I made:
1. In 'reportingunits.csv', I added the EPSG code to all rows of the column EPSGCode
2. In 'reportingunits.csv', I added a dummy WKT to the Geometry filed 
3. In 'aggregatedamounts.csv', I added a new column called "PowerType". This is due to the recent change by DPL
4. In the SQL database, I modified the CVs.State table to add a row with WY as state. This is strange problem because both the WaDE CV website and the spreadsheet in Google do have WY as one of the controlled vocabulary state. The SQL db CVs.State is not automatically updating itself. 
5. Add beneficial use category items to the CV database 

The import of the 'aggregatedamounts.csv' fails an error: "Cross group not valid". 
The full error text can be found in the file "errTest.txt" and the SQL procedure that catches the error in "aggamountSQL.sql".